Multiple fixes for the R2R code generator
As Simon discovered, one of my recent refactorings
of checks for the Marked flag from earlier this week
effectively broke import signatures; the biggest
downside of this fact was that CoreRun silently
switched over to JIT to that many bugs in the R2R
compiler went unnoticed.

By now I have fixed enough bugs to make
WriteLine("Hello world!") work again but nothing
beyond that. I plan to proceed bringing
up the "actual R2R functionality" for some of the
other features that got implemented but not actually
tested like my initial attempts at generics, after
that I'll start working on integration with Simon's
JIT interface change.

1) In VertexArray, the implementation was incorrect
as I originally misunderstood the Section.Pop concept
and I messed up some parentheses.

2) We weren't actually "rooting" the
RvaEmbeddedPointerIndirectionNode anywhere so it got
never Marked, that was causing the signature issues.
I have changed Import to point at it.

3) In ImportSectionsTableNode I had to put back the
handwritten element data collection because we need
to skip the empty import sections and properly assign
indices only to the non-skipped ones, otherwise
table indices in method precode fixups are incorrect.

4) I found out that the latest versions of the
produced R2R exe crash R2RDump. This is because I didn't
generate the AMD64 UNWIND_INFO structure in front
of the GC info. I thought I found a way to encode it
but apparently it's the wrong encoding. I have hotfixed
this for now by putting in an DWORD(0) which means
empty UNWIND_INFO.

Thanks

Tomas

+++ src/ILCompiler.ReadyToRun/src/Compiler/DependencyAnalysis/ReadyToRun/StringImportSignature.cs
@@ -25,7 +25,7 @@ namespace ILCompiler.DependencyAnalysis.ReadyToRun
             dataBuilder.AddSymbol(this);
 
             dataBuilder.EmitByte((byte)ReadyToRunFixupKind.READYTORUN_FIXUP_StringHandle);
-            SignatureBuilder.EmitData(ref dataBuilder, (uint)SignatureBuilder.RidFromToken((int)_token));
+            SignatureBuilder.EmitData(ref dataBuilder, (uint)SignatureBuilder.RidFromToken(_token));
 
             return dataBuilder.ToObjectData();
         }
